core+tests: rename remaining mentions of shamir to slip39
