+++ core/src/trezor/ui/slip39.py
