+++ tests/device_tests/test_msg_resetdevice_slip39_advanced.py
@@ -31,7 +31,7 @@ EXTERNAL_ENTROPY = b"zlutoucky kun upel divoke ody" * 2
 class TestMsgResetDeviceT2:
     # TODO: test with different options
     @pytest.mark.setup_client(uninitialized=True)
-    def test_reset_device_supershamir(self, client):
+    def test_reset_device_slip39_advanced(self, client):
         strength = 128
         word_count = 20
         member_threshold = 3
